+++ backend/core/src/listings/listings.service.ts
@@ -137,10 +137,11 @@ export class ListingsService {
   async create(listingDto: ListingCreateDto) {
     const listing = this.listingRepository.create({
       ...listingDto,
+      publishedAt: listingDto.status === ListingStatus.active ?? new Date(),
+      closedAt: listingDto.status === ListingStatus.closed ?? new Date(),
       property: plainToClass(PropertyCreateDto, listingDto),
     })
-    const saveResult = await listing.save()
-    return saveResult
+    return await listing.save()
   }
 
   async update(listingDto: ListingUpdateDto) {
@@ -160,9 +161,18 @@ export class ListingsService {
         availableUnits++
       }
     })
+
     listingDto.unitsAvailable = availableUnits
     Object.assign(listing, {
       ...plainToClass(Listing, listingDto, { excludeExtraneousValues: true }),
+      publishedAt:
+        listing.status !== ListingStatus.active && listingDto.status === ListingStatus.active
+          ? new Date()
+          : listing.publishedAt,
+      closedAt:
+        listing.status !== ListingStatus.closed && listingDto.status === ListingStatus.closed
+          ? new Date()
+          : listing.closedAt,
       property: plainToClass(
         PropertyUpdateDto,
         {
@@ -175,6 +185,7 @@ export class ListingsService {
         { excludeExtraneousValues: true }
       ),
     })
+
     return await this.listingRepository.save(listing)
   }
 
@@ -226,6 +237,7 @@ export class ListingsService {
       .getMany()
     for (const listing of listings) {
       listing.status = ListingStatus.closed
+      listing.closedAt = new Date()
     }
 
     await this.listingRepository.save(listings)